move productFixtures.ts out of /test/data so we don't add unnecessary test JSON to bundle
