+++ packages/components/src/test/data/constants.ts
@@ -20,11 +20,6 @@ import { AssayWizardModel } from '../../internal/components/assay/AssayWizardMod
 
 import { DELETE_SAMPLES_TOPIC } from '../../internal/util/helpLinks';
 
-import {
-    FREEZER_MANAGER_APP_PROPERTIES,
-    ProductFeature,
-    SAMPLE_MANAGER_APP_PROPERTIES,
-} from '../../internal/app/constants';
 import { AssayDefinitionModel, AssayDomainTypes } from '../../internal/AssayDefinitionModel';
 import { QueryInfo } from '../../public/QueryInfo';
 import { IFile } from '../../internal/components/files/models';
@@ -452,49 +447,4 @@ export const TestTypeDataType: EntityDataType = {
 export const TestTypeDataTypeWithEntityFilter: EntityDataType = {
     ...TestTypeDataType,
     filterArray: [Filter.create('Category', 'Source')],
-};
-
-export const TEST_LKSM_PROFESSIONAL_MODULE_CONTEXT = {
-    api: {
-        moduleNames: ['samplemanagement', 'inventory', 'assay', 'labbook'],
-    },
-    samplemanagement: {
-        productId: SAMPLE_MANAGER_APP_PROPERTIES.productId,
-    },
-    inventory: {
-        productId: FREEZER_MANAGER_APP_PROPERTIES.productId,
-    },
-    core: {
-        productFeatures: [ProductFeature.Workflow, ProductFeature.ELN, ProductFeature.Assay],
-    },
-};
-
-export const TEST_LKSM_STARTER_MODULE_CONTEXT = {
-    api: {
-        moduleNames: ['samplemanagement', 'inventory'],
-    },
-    samplemanagement: {
-        productId: SAMPLE_MANAGER_APP_PROPERTIES.productId,
-    },
-    inventory: {
-        productId: FREEZER_MANAGER_APP_PROPERTIES.productId,
-    },
-    core: {
-        productFeatures: [],
-    },
-};
-
-export const TEST_LKS_STARTER_MODULE_CONTEXT = {
-    api: {
-        moduleNames: ['samplemanagement', 'inventory', 'assay', 'premium'],
-    },
-    samplemanagement: {
-        productId: SAMPLE_MANAGER_APP_PROPERTIES.productId,
-    },
-    inventory: {
-        productId: FREEZER_MANAGER_APP_PROPERTIES.productId,
-    },
-    core: {
-        productFeatures: [ProductFeature.Assay],
-    },
 };
